Move binding and typing inside ExpressionNode
